AI_Engine/llm/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
numpy
requests
google-genai
sqlalchemy>=2.0
sqlalchemy-utils
sentence-transformers
